refactor ConnectButton
