+++ src/components/v2/Layout/Header/CoinInfo.tsx
@@ -4,6 +4,8 @@ import BigNumber from 'bignumber.js';
 import commaNumber from 'comma-number';
 import Typography from '@mui/material/Typography';
 import { Paper } from '@mui/material';
+import { useVaiUser } from 'hooks/useVaiUser';
+import { useMarketsUser } from 'hooks/useMarketsUser';
 import { Icon } from '../../Icon';
 import { useStyles } from './styles';
 
@@ -20,4 +22,14 @@ export const CoinInfo = ({ balance, coin }: { balance: BigNumber; coin: 'xvs' | 
       </Typography>
     </Paper>
   );
+};
+
+export const XvsCoinInfo = () => {
+  const { userVaiBalance } = useVaiUser();
+  return <CoinInfo balance={userVaiBalance} coin="vai" />;
+};
+
+export const VaiCoinInfo = () => {
+  const { userXvsBalance } = useMarketsUser();
+  return <CoinInfo balance={userXvsBalance} coin="xvs" />;
 };
