+++ src/components/v2/Layout/Header/index.tsx
@@ -4,8 +4,8 @@ import Typography from '@mui/material/Typography';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import { Toolbar } from '../Toolbar';
-import ConnectButton from '../../../Basic/ConnectButton';
-import { XvsCoinInfo, VaiCoinInfo } from './CoinInfoContainers';
+import ConnectButton from './ConnectButton';
+import { XvsCoinInfo, VaiCoinInfo } from './CoinInfo';
 import { useStyles } from './styles';
 
 interface IHeaderProps {
